Moved pod specific implementation to nativkubernetes
