+++ pkg/kubernetesutils/nativekubernetes/containers.go
@@ -1,4 +1,4 @@
-package nativekubernetesoo
+package nativekubernetes
 
 import (
 	"bytes"
