+++ pkg/kubernetesutils/nativekubernetes/pods.go
@@ -1,0 +1,260 @@
+package nativekubernetes
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	"github.com/asciich/asciichgolangpublic/pkg/commandexecutor/commandoutput"
+	"github.com/asciich/asciichgolangpublic/pkg/kubernetesutils/kubernetesparameteroptions"
+	"github.com/asciich/asciichgolangpublic/pkg/logging"
+	"github.com/asciich/asciichgolangpublic/pkg/tracederrors"
+	corev1 "k8s.io/api/core/v1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/rest"
+)
+
+func DeletePod(ctx context.Context, clientset *kubernetes.Clientset, podName string, namespace string) error {
+	if clientset == nil {
+		return tracederrors.TracedErrorNil("clientSet")
+	}
+
+	if podName == "" {
+		return tracederrors.TracedErrorEmptyString("podName")
+	}
+
+	if namespace == "" {
+		return tracederrors.TracedErrorEmptyString("namespace")
+	}
+
+	logging.LogInfoByCtxf(ctx, "Delete pod '%s' in namepace '%s' started.", podName, namespace)
+
+	deletePolicy := metav1.DeletePropagationBackground
+	err := clientset.CoreV1().Pods(namespace).Delete(ctx, podName, metav1.DeleteOptions{
+		PropagationPolicy: &deletePolicy,
+	})
+	if err == nil {
+		logging.LogChangedByCtxf(ctx, "Pod '%s' in namepsace '%s' deleted.", podName, namespace)
+	} else {
+		if apierrors.IsNotFound(err) {
+			logging.LogInfoByCtxf(ctx, "Pod '%s' already absent in namespace '%s'.", podName, namespace)
+		} else {
+			return tracederrors.TracedErrorf("Failed to delete pod '%s' in namespace '%s': %w", podName, namespace, err)
+		}
+	}
+
+	logging.LogInfoByCtxf(ctx, "Delete pod '%s' in namepace '%s' finished.", podName, namespace)
+
+	return nil
+}
+
+func RunCommandInTemporaryPod(ctx context.Context, config *rest.Config, options *kubernetesparameteroptions.RunCommandOptions) (*commandoutput.CommandOutput, error) {
+	if config == nil {
+		return nil, tracederrors.TracedErrorNil("config")
+	}
+
+	if options == nil {
+		return nil, tracederrors.TracedErrorNil("options")
+	}
+
+	namespace, err := options.GetNamespaceName()
+	if err != nil {
+		return nil, err
+	}
+
+	podName, err := options.GetPodName()
+	if err != nil {
+		return nil, err
+	}
+
+	imageName, err := options.GetImageName()
+	if err != nil {
+		return nil, err
+	}
+
+	command, err := options.GetCommand()
+	if err != nil {
+		return nil, err
+	}
+
+	logging.LogInfoByCtxf(ctx, "Run command in temporary pod '%s' in namespace '%s' using container image '%s' started.", podName, namespace, imageName)
+
+	containerName := podName
+	pod := &corev1.Pod{
+		ObjectMeta: metav1.ObjectMeta{
+			Name: podName,
+		},
+		Spec: corev1.PodSpec{
+			Containers: []corev1.Container{
+				{
+					Name:    containerName,
+					Image:   imageName,
+					Command: command,
+					Stdin:   true,
+					TTY:     true,
+				},
+			},
+			RestartPolicy: corev1.RestartPolicyNever,
+		},
+	}
+
+	clientset, err := kubernetes.NewForConfig(config)
+	if err != nil {
+		return nil, tracederrors.TracedErrorf("Unable to get client set: %w", err)
+	}
+
+	logging.LogInfoByCtxf(ctx, "Going to start pod '%s' in namespace '%s' using container image '%s'.", podName, namespace, imageName)
+	_, err = clientset.CoreV1().Pods(namespace).Create(ctx, pod, metav1.CreateOptions{})
+	if err != nil {
+		if apierrors.IsAlreadyExists(err) && options.DeleteAlreadyExistingPod {
+			logging.LogInfoByCtxf(ctx, "Going to delete pod already existing pod '%s' in namespace '%s' before running command.", podName, namespace)
+			err = DeletePod(ctx, clientset, podName, namespace)
+			if err != nil {
+				return nil, err
+			}
+			_, err = clientset.CoreV1().Pods(namespace).Create(ctx, pod, metav1.CreateOptions{})
+			if err != nil {
+				return nil, tracederrors.TracedErrorf("Error creating Pod: %w", err)
+			}
+		} else {
+			return nil, tracederrors.TracedErrorf("Error creating Pod: %w", err)
+		}
+	}
+
+	// Ensure pod is deleted after executing the command
+	defer func() {
+		_ = DeletePod(ctx, clientset, podName, namespace)
+	}()
+
+	err = waitForPodSucceeded(ctx, clientset, namespace, podName, time.Minute*1)
+	if err != nil {
+		return nil, err
+	}
+
+	stdout, stderr, err := GetContainerLogs(ctx, clientset, namespace, podName, containerName)
+	if err != nil {
+		return nil, err
+	}
+
+	var retVal = 0
+	output := &commandoutput.CommandOutput{
+		ReturnCode: &retVal,
+		Stdout:     &stdout,
+		Stderr:     &stderr,
+	}
+
+	logging.LogInfoByCtxf(ctx, "Run command in temporary pod '%s' in namespace '%s' using container image '%s' finished.", podName, namespace, imageName)
+
+	return output, nil
+}
+
+func waitForPodRunning(ctx context.Context, clientSet *kubernetes.Clientset, namespace string, podName string, timeout time.Duration) error {
+	if clientSet == nil {
+		return tracederrors.TracedErrorNil("clientSet")
+	}
+
+	if podName == "" {
+		return tracederrors.TracedErrorEmptyString("podName")
+	}
+
+	if namespace == "" {
+		return tracederrors.TracedErrorEmptyString("namespace")
+	}
+
+	logging.LogInfoByCtxf(ctx, "Wait for pod '%s' in namespace '%s' to be running started.", podName, namespace)
+
+	w, err := clientSet.CoreV1().Pods(namespace).Watch(ctx, metav1.ListOptions{
+		FieldSelector: fmt.Sprintf("metadata.name=%s", podName),
+	})
+	if err != nil {
+		return fmt.Errorf("failed to set up watch for pod %s: %w", podName, err)
+	}
+	defer w.Stop()
+
+	timer := time.NewTimer(timeout)
+	defer timer.Stop()
+
+	for {
+		select {
+		case event, ok := <-w.ResultChan():
+			if !ok {
+				return tracederrors.TracedErrorf("watch channel closed unexpectedly when waiting for pod '%s' in namespace '%s' to be running", podName, namespace)
+			}
+
+			pod, ok := event.Object.(*corev1.Pod)
+			if !ok {
+				continue
+			}
+
+			if pod.Status.Phase == corev1.PodRunning {
+				logging.LogInfoByCtxf(ctx, "Wait for pod '%s' in namespace '%s' to be running finished. Pod is now running", podName, namespace)
+				return nil
+			}
+			// If pod is in a failed state, exit early
+			if pod.Status.Phase == corev1.PodFailed || pod.Status.Phase == corev1.PodSucceeded {
+				return tracederrors.TracedErrorf("pod '%s' in namespace '%s' entered phase %s before running", podName, namespace, pod.Status.Phase)
+			}
+		case <-timer.C:
+			return tracederrors.TracedErrorf("timeout waiting for pod '%s' in namespace '%s' to be running", podName, namespace)
+		case <-ctx.Done():
+			return ctx.Err() // Context was cancelled
+		}
+	}
+
+}
+
+func waitForPodSucceeded(ctx context.Context, clientSet *kubernetes.Clientset, namespace string, podName string, timeout time.Duration) error {
+	if clientSet == nil {
+		return tracederrors.TracedErrorNil("clientSet")
+	}
+
+	if podName == "" {
+		return tracederrors.TracedErrorEmptyString("podName")
+	}
+
+	if namespace == "" {
+		return tracederrors.TracedErrorEmptyString("namespace")
+	}
+
+	logging.LogInfoByCtxf(ctx, "Wait for pod '%s' in namespace '%s' to be succeeded started.", podName, namespace)
+
+	w, err := clientSet.CoreV1().Pods(namespace).Watch(ctx, metav1.ListOptions{
+		FieldSelector: fmt.Sprintf("metadata.name=%s", podName),
+	})
+	if err != nil {
+		return fmt.Errorf("failed to set up watch for pod %s: %w", podName, err)
+	}
+	defer w.Stop()
+
+	timer := time.NewTimer(timeout)
+	defer timer.Stop()
+
+	for {
+		select {
+		case event, ok := <-w.ResultChan():
+			if !ok {
+				return tracederrors.TracedErrorf("watch channel closed unexpectedly when waiting for pod '%s' in namespace '%s' to be succeeded", podName, namespace)
+			}
+
+			pod, ok := event.Object.(*corev1.Pod)
+			if !ok {
+				continue
+			}
+
+			if pod.Status.Phase == corev1.PodSucceeded {
+				logging.LogInfoByCtxf(ctx, "Wait for pod '%s' in namespace '%s' to be succeeded finished. Pod is now succeeded.", podName, namespace)
+				return nil
+			}
+			// If pod is in a failed state, exit early
+			if pod.Status.Phase == corev1.PodFailed {
+				return tracederrors.TracedErrorf("pod '%s' in namespace '%s' failed", podName, namespace)
+			}
+		case <-timer.C:
+			return tracederrors.TracedErrorf("timeout waiting for pod '%s' in namespace '%s' to be succeeded", podName, namespace)
+		case <-ctx.Done():
+			return ctx.Err() // Context was cancelled
+		}
+	}
+}
