--- pkg/kubernetesutils/nativekubernetesoo/Pods.go
@@ -1,1 +1,0 @@
-package nativekubernetesoo
